move generator tests into tests module
they test not just the generator, but parts of the lock file extraction etc as well, so they are more like integration tests

not putting them into the top level tests/ because we are reserving that location for real end-to-end integration tests using libcnb, without access to modules in the crate
