+++ buildpacks/php/src/platform/generator.rs
@@ -313,618 +313,3 @@ pub(crate) fn generate_platform_json(
         ..Default::default()
     })
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use crate::package_manager::composer::PlatformExtractorNotice;
-
-    use assert_json_diff::{assert_json_matches_no_panic, CompareMode, Config};
-    use figment::providers::{Format, Serialized, Toml};
-    use figment::{value::magic::RelativePathBuf, Figment};
-    use serde::{Deserialize, Serialize};
-    use serde_json::{Map, Value};
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-    use std::{env, fs};
-
-    #[derive(Deserialize, Serialize)]
-    struct ComposerLockTestCaseConfig {
-        name: Option<String>,
-        description: Option<String>,
-        lock: Option<RelativePathBuf>,
-        stack: String,
-        expected_result: Option<RelativePathBuf>,
-        expected_extractor_notices: Option<Vec<String>>, // TODO: can we use PlatformExtractorNotice?
-        expected_finalizer_notices: Option<Vec<String>>, // TODO: can we use PlatformFinalizerNotice?
-        expect_extractor_failure: Option<String>,        // TODO: can we use PlatformExtractorError?
-        expect_generator_failure: Option<String>,        // TODO: can we use PlatformGeneratorError?
-        expect_finalizer_failure: Option<String>,        // TODO: can we use PlatformFinalizerError?
-        install_dev: bool,
-        repositories: Vec<Url>,
-    }
-
-    impl Default for ComposerLockTestCaseConfig {
-        fn default() -> Self {
-            let stack = "heroku-20";
-            Self {
-                name: None,
-                description: None,
-                lock: None,
-                stack: stack.to_string(),
-                expected_result: None,
-                expected_extractor_notices: None,
-                expected_finalizer_notices: None,
-                expect_generator_failure: None,
-                expect_extractor_failure: None,
-                expect_finalizer_failure: None,
-                install_dev: false,
-                repositories: vec![Url::parse(&format!(
-                    "https://lang-php.s3.us-east-1.amazonaws.com/dist-{stack}-cnb/packages.json",
-                ))
-                .unwrap()],
-            }
-        }
-    }
-
-    impl<P: AsRef<Path>> From<P> for ComposerLockTestCaseConfig {
-        fn from(p: P) -> Self {
-            let dir = p.as_ref();
-            let lock = dir.join("composer.lock");
-            let expected_result = dir.join("expected_platform_composer.json");
-            Self {
-                name: Some(dir.file_name().unwrap().to_string_lossy().to_string()),
-                lock: lock.try_exists().unwrap().then_some(lock.into()),
-                expected_result: expected_result
-                    .try_exists()
-                    .unwrap()
-                    .then_some(expected_result.into()),
-                ..Default::default()
-            }
-        }
-    }
-
-    #[test]
-    fn make_platform_json_with_fixtures() {
-        let installer_path = &PathBuf::from("../../support/installer");
-
-        fs::read_dir(Path::new("tests/fixtures/platform/generator"))
-            .unwrap()
-            .filter(|der| der.as_ref().unwrap().metadata().unwrap().is_dir())
-            .filter_map(|der| {
-                let p = der.unwrap().path();
-                // merge our auto-built config (from Path) and a config.toml, if it exists
-                let case: ComposerLockTestCaseConfig =
-                    Figment::from(Serialized::defaults(ComposerLockTestCaseConfig::from(&p)))
-                        .merge(Toml::file(p.join("config.toml")))
-                        .extract()
-                        .unwrap();
-                // skip if there isn't even a lock file in the dir
-                case.lock.is_some().then_some(case)
-            })
-            .for_each(|case| {
-                let lock = serde_json::from_str(
-                    // .relative() will allow specifying the file name in the config
-                    &fs::read_to_string(case.lock.as_ref().unwrap().relative()).unwrap(),
-                )
-                .unwrap();
-
-                // FIRST: from the lock file, extract a generator config and packages list
-
-                let generator_input = package_manager::composer::extract_from_lock(&lock);
-
-                // first check: was this even supposed to succeed or fail?
-                assert_eq!(
-                    generator_input.is_ok(),
-                    case.expect_extractor_failure.is_none(),
-                    "case {}: lock extraction expected to {}, but it didn't",
-                    case.name.as_ref().unwrap(),
-                    if generator_input
-                        .is_ok() { "fail" } else { "succeed" },
-                );
-
-                // on failure, check if the type of failure what was the test expected
-                let mut extractor_notices  = Vec::<PlatformExtractorNotice>::new();
-                let generator_input  = match generator_input
-                {
-                    Ok(v) => v.unwrap(&mut extractor_notices),
-                    Err(e) => {
-                        assert!(
-                            case.expect_extractor_failure.is_some(),
-                            "case {}: lock extraction failed, but config has no expect_extractor_failure type specified",
-                            case.name.as_ref().unwrap()
-                        );
-
-                        assert_eq!(
-                            format!("{e:?}"),
-                            case.expect_extractor_failure.unwrap(),
-                            "case {}: lock extraction failed as expected, but with mismatched failure type",
-                            case.name.as_ref().unwrap()
-                        );
-
-                        return;
-                    }
-                };
-
-                // fetch all notices and compare them against the list of expected notices
-                assert_eq!(
-                    extractor_notices
-                        .iter()
-                        .map(|v| format!("{v:?}"))
-                        .collect::<HashSet<String>>(),
-                    case.expected_extractor_notices
-                        .unwrap_or_default()
-                        .into_iter()
-                        .collect::<HashSet<String>>(),
-                    "case {}: mismatched lock extractor notices (left = generated, right = expected)",
-                    case.name.as_ref().unwrap()
-                );
-
-                // SECOND: generate "platform.json" from the extracted config and packages list
-
-                let generated_json_package = generate_platform_json(
-                    &generator_input,
-                    &case.stack,
-                    installer_path,
-                    &case.repositories,
-                );
-
-                // first check: was this even supposed to succeed or fail?
-                assert_eq!(
-                    generated_json_package.is_ok(),
-                    case.expect_generator_failure.is_none(),
-                    "case {}: generation expected to {}, but it didn't",
-                    case.name.as_ref().unwrap(),
-                    if generated_json_package
-                        .is_ok() { "fail" } else { "succeed" },
-                );
-
-                // on failure, check if the type of failure what was the test expected
-                let mut generated_json_package = match generated_json_package
-                {
-                    Ok(v) => v,
-                    Err(e) => {
-                        assert!(
-                            case.expect_generator_failure.is_some(),
-                            "case {}: generation failed, but config has no expect_generator_failure type specified",
-                            case.name.as_ref().unwrap()
-                        );
-
-                        assert_eq!(
-                            format!("{e:?}"),
-                            case.expect_generator_failure.unwrap(),
-                            "case {}: generation failed as expected, but with mismatched failure type",
-                            case.name.as_ref().unwrap()
-                        );
-
-                        return;
-                    }
-                };
-
-                // THIRD: post-process the generated result to ensure/validate runtime requirements etc
-                let ensure_runtime_requirement_result = package_manager::composer::ensure_runtime_requirement(&mut generated_json_package);
-
-                // first check: was this even supposed to succeed or fail?
-                assert_eq!(
-                    ensure_runtime_requirement_result.is_ok(),
-                    case.expect_finalizer_failure.is_none(),
-                    "case {}: finalizing expected to {}, but it didn't",
-                    case.name.as_ref().unwrap(),
-                    if ensure_runtime_requirement_result
-                        .is_ok() { "fail" } else { "succeed" },
-                );
-
-                // on failure, check if the type of failure what was the test expected
-                let finalizer_notices = match ensure_runtime_requirement_result
-                {
-                    Ok(v) => v,
-                    Err(e) => {
-                        assert!(
-                            case.expect_finalizer_failure.is_some(),
-                            "case {}: finalizing failed, but config has no expect_finalizer_failure type specified",
-                            case.name.as_ref().unwrap()
-                        );
-
-                        assert_eq!(
-                            format!("{e:?}"),
-                            case.expect_finalizer_failure.unwrap(),
-                            "case {}: finalizing failed as expected, but with mismatched failure type",
-                            case.name.as_ref().unwrap()
-                        );
-
-                        return;
-                    }
-                };
-
-                // fetch all notices and compare them against the list of expected notices
-                assert_eq!(
-                    finalizer_notices
-                        .iter()
-                        .map(|v| format!("{v:?}"))
-                        .collect::<HashSet<String>>(),
-                    case.expected_finalizer_notices
-                        .unwrap_or_default()
-                        .into_iter()
-                        .collect::<HashSet<String>>(),
-                    "case {}: mismatched finalizer notices (left = generated, right = expected)",
-                    case.name.as_ref().unwrap()
-                );
-
-                if !case.install_dev {
-                    // remove require-dev if we do not want dev installs
-                    generated_json_package.package.require_dev.take();
-                }
-
-                let mut expected_json_object: Map<String, Value> = serde_json::from_str(
-                    &fs::read_to_string(case.expected_result.unwrap().relative()).unwrap(),
-                )
-                .unwrap();
-
-                let generated_json_value =
-                    serde_json::value::to_value(&generated_json_package).unwrap();
-                let generated_json_object = generated_json_value.as_object().unwrap();
-
-                let generated_keys: HashSet<String> =
-                    generated_json_object.keys().cloned().collect();
-                let expected_keys: HashSet<String> = expected_json_object.keys().cloned().collect();
-
-                // check if all of the expected keys are there (and only those)
-                assert_eq!(
-                    &generated_keys,
-                    &expected_keys,
-                    "case {}: mismatched keys (left = generated, right = expected)",
-                    case.name.as_ref().unwrap()
-                );
-
-                // validate each key in the generated JSON
-                // we have to do this because we want to treat e.g. the "provide" key a bit differently
-                for key in expected_keys {
-                    let generated_value = generated_json_object.get(key.as_str()).unwrap();
-                    let expected_value = match key.as_str() {
-                        k @ "provide" => {
-                            if let Value::Object(obj) =
-                                &mut expected_json_object.get_mut(k).unwrap()
-                            {
-                                // for heroku-sys/heroku, we want to check that the generated value starts with the expected value
-                                // (since the version strings are like XX.YYYY.MM.DD, with XX being the stack version number)
-                                obj.entry("heroku-sys/heroku").and_modify(|exp| {
-                                    let gen = generated_value.get("heroku-sys/heroku").unwrap();
-                                    if gen.as_str().unwrap().starts_with(exp.as_str().unwrap()) {
-                                        *exp = gen.clone();
-                                    }
-                                });
-                            }
-                            expected_json_object.get(k).unwrap()
-                        }
-                        // k @ "repositories" => expected_json_object.get(k).unwrap(), // maybe normalize plugin repo path, maybe sort packages in "package" repo?
-                        k => expected_json_object.get(k).unwrap(),
-                    };
-
-                    let comparison = assert_json_matches_no_panic(
-                        generated_value,
-                        expected_value,
-                        Config::new(CompareMode::Strict),
-                    )
-                    .map_err(|err| {
-                        format!("case {}, key {}: {}", case.name.as_ref().unwrap(), key, err)
-                    });
-
-                    assert!(comparison.is_ok(), "{}", comparison.unwrap_err());
-                }
-            });
-    }
-
-    // #[test]
-    fn yo() {
-        let composer_lock = r#"{
-    "_readme": [
-        "This file locks the dependencies of your project to a known state",
-        "Read more about it at https://getcomposer.org/doc/01-basic-usage.md#installing-dependencies",
-        "This file is @generated automatically"
-    ],
-    "content-hash": "c2b9dcae256d1b255b7265eef089f6c3",
-    "packages": [
-        {
-            "name": "symfony/polyfill-php80",
-            "version": "v1.23.1",
-            "source": {
-                "type": "git",
-                "url": "https://github.com/symfony/polyfill-php80.git",
-                "reference": "1100343ed1a92e3a38f9ae122fc0eb21602547be"
-            },
-            "dist": {
-                "type": "zip",
-                "url": "https://api.github.com/repos/symfony/polyfill-php80/zipball/1100343ed1a92e3a38f9ae122fc0eb21602547be",
-                "reference": "1100343ed1a92e3a38f9ae122fc0eb21602547be",
-                "shasum": ""
-            },
-            "require": {
-                "php": ">=7.1"
-            },
-            "type": "library",
-            "extra": {
-                "branch-alias": {
-                    "dev-main": "1.23-dev"
-                },
-                "thanks": {
-                    "name": "symfony/polyfill",
-                    "url": "https://github.com/symfony/polyfill"
-                }
-            },
-            "autoload": {
-                "psr-4": {
-                    "Symfony\\Polyfill\\Php80\\": ""
-                },
-                "files": [
-                    "bootstrap.php"
-                ],
-                "classmap": [
-                    "Resources/stubs"
-                ]
-            },
-            "notification-url": "https://packagist.org/downloads/",
-            "license": [
-                "MIT"
-            ],
-            "authors": [
-                {
-                    "name": "Ion Bazan",
-                    "email": "ion.bazan@gmail.com"
-                },
-                {
-                    "name": "Nicolas Grekas",
-                    "email": "p@tchwork.com"
-                },
-                {
-                    "name": "Symfony Community",
-                    "homepage": "https://symfony.com/contributors"
-                }
-            ],
-            "description": "Symfony polyfill backporting some PHP 8.0+ features to lower PHP versions",
-            "homepage": "https://symfony.com",
-            "keywords": [
-                "compatibility",
-                "polyfill",
-                "portable",
-                "shim"
-            ],
-            "support": {
-                "source": "https://github.com/symfony/polyfill-php80/tree/v1.23.1"
-            },
-            "funding": [
-                {
-                    "url": "https://symfony.com/sponsor",
-                    "type": "custom"
-                },
-                {
-                    "url": "https://github.com/fabpot",
-                    "type": "github"
-                },
-                {
-                    "url": "https://tidelift.com/funding/github/packagist/symfony/symfony",
-                    "type": "tidelift"
-                }
-            ],
-            "time": "2021-07-28T13:41:28+00:00"
-        },
-        {
-            "name": "symfony/process",
-            "version": "v5.1.0-RC1",
-            "source": {
-                "type": "git",
-                "url": "https://github.com/symfony/process.git",
-                "reference": "14c0d48567aafd6b24001866de32ae45b0e3e1d1"
-            },
-            "dist": {
-                "type": "zip",
-                "url": "https://api.github.com/repos/symfony/process/zipball/14c0d48567aafd6b24001866de32ae45b0e3e1d1",
-                "reference": "14c0d48567aafd6b24001866de32ae45b0e3e1d1",
-                "shasum": ""
-            },
-            "require": {
-                "php": "^7.2.5",
-                "symfony/polyfill-php80": "^1.15"
-            },
-            "type": "library",
-            "extra": {
-                "branch-alias": {
-                    "dev-master": "5.1-dev"
-                }
-            },
-            "autoload": {
-                "psr-4": {
-                    "Symfony\\Component\\Process\\": ""
-                },
-                "exclude-from-classmap": [
-                    "/Tests/"
-                ]
-            },
-            "notification-url": "https://packagist.org/downloads/",
-            "license": [
-                "MIT"
-            ],
-            "authors": [
-                {
-                    "name": "Fabien Potencier",
-                    "email": "fabien@symfony.com"
-                },
-                {
-                    "name": "Symfony Community",
-                    "homepage": "https://symfony.com/contributors"
-                }
-            ],
-            "description": "Symfony Process Component",
-            "homepage": "https://symfony.com",
-            "support": {
-                "source": "https://github.com/symfony/process/tree/master"
-            },
-            "funding": [
-                {
-                    "url": "https://symfony.com/sponsor",
-                    "type": "custom"
-                },
-                {
-                    "url": "https://github.com/fabpot",
-                    "type": "github"
-                },
-                {
-                    "url": "https://tidelift.com/funding/github/packagist/symfony/symfony",
-                    "type": "tidelift"
-                }
-            ],
-            "time": "2020-04-15T16:09:08+00:00"
-        }
-    ],
-    "packages-dev": [
-        {
-            "name": "kahlan/kahlan",
-            "version": "5.1.3",
-            "source": {
-                "type": "git",
-                "url": "https://github.com/kahlan/kahlan.git",
-                "reference": "bbf99064b7b78049f58e20138bee18fcdee3573e"
-            },
-            "dist": {
-                "type": "zip",
-                "url": "https://api.github.com/repos/kahlan/kahlan/zipball/bbf99064b7b78049f58e20138bee18fcdee3573e",
-                "reference": "bbf99064b7b78049f58e20138bee18fcdee3573e",
-                "shasum": ""
-            },
-            "require": {
-                "php": ">=7.1"
-            },
-            "require-dev": {
-                "squizlabs/php_codesniffer": "^3.4"
-            },
-            "bin": [
-                "bin/kahlan"
-            ],
-            "type": "library",
-            "autoload": {
-                "psr-4": {
-                    "Kahlan\\": "src/"
-                },
-                "files": [
-                    "src/functions.php"
-                ]
-            },
-            "notification-url": "https://packagist.org/downloads/",
-            "license": [
-                "MIT"
-            ],
-            "authors": [
-                {
-                    "name": "CrysaLEAD"
-                }
-            ],
-            "description": "The PHP Test Framework for Freedom, Truth and Justice.",
-            "keywords": [
-                "BDD",
-                "Behavior-Driven Development",
-                "Monkey Patching",
-                "TDD",
-                "mock",
-                "stub",
-                "testing",
-                "unit test"
-            ],
-            "support": {
-                "issues": "https://github.com/kahlan/kahlan/issues",
-                "source": "https://github.com/kahlan/kahlan/tree/5.1.3"
-            },
-            "time": "2021-06-13T11:14:50+00:00"
-        }
-    ],
-    "aliases": [],
-    "minimum-stability": "RC",
-    "stability-flags": {
-        "symfony/process": 5
-    },    
-    "prefer-stable": true,
-    "prefer-lowest": false,
-    "platform": {
-        "ext-gmp": "*",
-        "ext-intl": "*",
-        "ext-mbstring": "*",
-        "ext-redis": "*",
-        "ext-sqlite3": "*",
-        "ext-ldap": "*",
-        "ext-imap": "*",
-        "ext-blackfire": "*"
-    },
-    "platform-dev": {
-        "ext-pcov": "*"
-    },
-    "plugin-api-version": "2.3.0"
-}
-"#;
-        let l: ComposerLock = serde_json::from_str(composer_lock).unwrap();
-
-        let stack = env::var("STACK").unwrap_or_else(|_| "heroku-22".to_string());
-
-        // our default repo
-        let default_repos = vec![Url::parse(
-            format!("https://lang-php.s3.us-east-1.amazonaws.com/dist-{stack}-cnb/",).as_str(),
-        )
-        .unwrap()];
-        // anything user-supplied
-        let byo_repos = env::var("HEROKU_PHP_PLATFORM_REPOSITORIES").unwrap_or_default();
-        let all_repos = crate::platform::platform_repository_urls_from_defaults_and_list(
-            &default_repos,
-            byo_repos,
-        )
-        .unwrap();
-
-        let generator_input = package_manager::composer::extract_from_lock(&l)
-            .unwrap()
-            .value;
-
-        let pj = serde_json::to_string_pretty(
-            &generate_platform_json(
-                &generator_input,
-                &stack,
-                &PathBuf::from("../../support/installer"),
-                &all_repos,
-            )
-            .unwrap(), // FIXME: handle
-        )
-        .unwrap();
-
-        println!("{pj}");
-    }
-
-    // #[test]
-    fn nothing() {
-        let stack = env::var("STACK").unwrap_or_else(|_| "heroku-22".to_string());
-
-        // our default repo
-        let default_repos = vec![Url::parse(
-            format!("https://lang-php.s3.us-east-1.amazonaws.com/dist-{stack}-cnb/",).as_str(),
-        )
-        .unwrap()];
-
-        let generator_input = PlatformJsonGeneratorInput {
-            additional_require: Some(HashMap::from([
-                ("heroku-sys/composer".to_string(), "*".to_string()),
-                ("heroku-sys/php".to_string(), "*".to_string()),
-            ])),
-            ..Default::default()
-        };
-
-        let mut pj = generate_platform_json(
-            &generator_input,
-            &stack,
-            &PathBuf::from("../../support/installer"),
-            &default_repos,
-        )
-        .unwrap();
-
-        // remove require-dev if we do not want dev installs
-        pj.package.require_dev.take();
-
-        let pj = serde_json::to_string_pretty(&pj).unwrap();
-
-        println!("{pj}");
-    }
-}
